+++ stars-core/src/test/kotlin/tools/aqua/stars/core/tsc/TSCNWayFeatureCombinationCountingTest.kt
@@ -20,6 +20,7 @@ package tools.aqua.stars.core.tsc
 import java.math.BigInteger
 import kotlin.test.Test
 import kotlin.test.assertEquals
+import kotlin.test.assertTrue
 import tools.aqua.stars.core.SimpleEntity
 import tools.aqua.stars.core.SimpleSegment
 import tools.aqua.stars.core.SimpleTickData
@@ -47,12 +48,16 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           all("root") {
-            leaf("leaf 1")
-            leaf("leaf 2")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
           }
         }
 
     assertEquals(BigInteger.TWO, testTSC.countAllPossibleNWayPredicateCombinations(1))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(1).all { it.elements.size == 1 },
+    )
+    assertEquals(2, testTSC.getAllPossibleNWayPredicateCombinations(1).size)
   }
 
   /** Test 1-way feature combination counting for optional node. */
@@ -61,12 +66,16 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           optional("root") {
-            leaf("leaf 1")
-            leaf("leaf 2")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
           }
         }
 
     assertEquals(BigInteger.TWO, testTSC.countAllPossibleNWayPredicateCombinations(1))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(1).all { it.elements.size == 1 },
+    )
+    assertEquals(2, testTSC.getAllPossibleNWayPredicateCombinations(1).size)
   }
 
   /** Test 1-way feature combination counting for any node. */
@@ -75,12 +84,16 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           any("root") {
-            leaf("leaf 1")
-            leaf("leaf 2")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
           }
         }
 
     assertEquals(BigInteger.TWO, testTSC.countAllPossibleNWayPredicateCombinations(1))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(1).all { it.elements.size == 1 },
+    )
+    assertEquals(2, testTSC.getAllPossibleNWayPredicateCombinations(1).size)
   }
 
   /** Test 1-way feature combination counting for exclusive node. */
@@ -89,12 +102,16 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           exclusive("root") {
-            leaf("leaf 1")
-            leaf("leaf 2")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
           }
         }
 
     assertEquals(BigInteger.TWO, testTSC.countAllPossibleNWayPredicateCombinations(1))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(1).all { it.elements.size == 1 },
+    )
+    assertEquals(2, testTSC.getAllPossibleNWayPredicateCombinations(1).size)
   }
 
   /** Test 2-way feature combination counting for bounded node. */
@@ -103,12 +120,16 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           bounded("root", 0 to 2) {
-            leaf("leaf 1")
-            leaf("leaf 2")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
           }
         }
 
     assertEquals(BigInteger.TWO, testTSC.countAllPossibleNWayPredicateCombinations(1))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(1).all { it.elements.size == 1 },
+    )
+    assertEquals(2, testTSC.getAllPossibleNWayPredicateCombinations(1).size)
   }
 
   /** Test 2-way feature combination counting for bounded node with bounds equals to 0. */
@@ -117,12 +138,112 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           bounded("root", 0 to 0) {
-            leaf("leaf 1")
-            leaf("leaf 2")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
           }
         }
 
     assertEquals(BigInteger.ZERO, testTSC.countAllPossibleNWayPredicateCombinations(1))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(1).all { it.elements.size == 1 },
+    )
+    assertEquals(0, testTSC.getAllPossibleNWayPredicateCombinations(1).size)
+  }
+
+  /** Test 1-way feature combination counting for with inner feature nodes. */
+  @Test
+  fun `Test 1-way feature combination counting for with inner feature nodes without condition`() {
+    val testTSC =
+        tsc<E, T, S, U, D> {
+          all("root") {
+            optional("optional 1") {
+              optional("optional 2") { leaf("leaf 1") { condition { true } } }
+            }
+          }
+        }
+
+    assertEquals(BigInteger.ONE, testTSC.countAllPossibleNWayPredicateCombinations(1))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(1).all { it.elements.size == 1 },
+    )
+    assertEquals(1, testTSC.getAllPossibleNWayPredicateCombinations(1).size)
+  }
+
+  /** Test 1-way feature combination counting with one inner feature node with a condition. */
+  @Test
+  fun `Test 1-way feature combination counting with one inner feature node with a condition`() {
+    val testTSC =
+        tsc<E, T, S, U, D> {
+          all("root") {
+            optional("optional 1") {
+              condition { true }
+              optional("optional 2") { leaf("leaf 1") { condition { true } } }
+            }
+          }
+        }
+
+    assertEquals(2, testTSC.getAllPossibleNWayPredicateCombinations(1).size)
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(1).all { it.elements.size == 1 },
+    )
+    assertEquals(BigInteger.TWO, testTSC.countAllPossibleNWayPredicateCombinations(1))
+  }
+
+  /** Test 1-way feature combination counting with two inner feature nodes with a condition. */
+  @Test
+  fun `Test 1-way feature combination counting with two inner feature nodes with a condition`() {
+    val testTSC =
+        tsc<E, T, S, U, D> {
+          all("root") {
+            optional("optional 1") {
+              condition { true }
+              optional("optional 2") {
+                condition { true }
+                leaf("leaf 1") { condition { true } }
+              }
+            }
+          }
+        }
+
+    assertEquals(3, testTSC.getAllPossibleNWayPredicateCombinations(1).size)
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(1).all { it.elements.size == 1 },
+    )
+    assertEquals(BigInteger("3"), testTSC.countAllPossibleNWayPredicateCombinations(1))
+  }
+
+  /** Test 1-way feature combination counting with two inner feature nodes with a condition. */
+  @Test
+  fun `Test 1-way feature combination counting with two inner feature nodes without a condition`() {
+    val testTSC =
+        tsc<E, T, S, U, D> {
+          all("root") {
+            optional("optional 1") {
+              optional("optional 2") { leaf("leaf 1") { condition { true } } }
+            }
+          }
+        }
+
+    assertEquals(1, testTSC.getAllPossibleNWayPredicateCombinations(1).size)
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(1).all { it.elements.size == 1 },
+    )
+    assertEquals(BigInteger("1"), testTSC.countAllPossibleNWayPredicateCombinations(1))
+  }
+
+  /** Test 1-way feature combination counting with all nodes without a condition. */
+  @Test
+  fun `Test 1-way feature combination counting with all nodes without a condition`() {
+    val testTSC =
+        tsc<E, T, S, U, D> {
+          all("root") { optional("optional 1") { optional("optional 2") { leaf("leaf 1") } } }
+        }
+
+    assertEquals(0, testTSC.getAllPossibleNWayPredicateCombinations(1).size)
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(1).all { it.elements.size == 1 },
+    )
+    assertEquals(BigInteger("0"), testTSC.countAllPossibleNWayPredicateCombinations(1))
   }
 
   // endregion
@@ -135,12 +256,16 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           all("root") {
-            leaf("leaf 1")
-            leaf("leaf 2")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
           }
         }
 
     assertEquals(BigInteger.ONE, testTSC.countAllPossibleNWayPredicateCombinations(2))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(2).all { it.elements.size == 2 },
+    )
+    assertEquals(1, testTSC.getAllPossibleNWayPredicateCombinations(2).size)
   }
 
   /** Test 2-way feature combination counting for optional node. */
@@ -149,12 +274,16 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           optional("root") {
-            leaf("leaf 1")
-            leaf("leaf 2")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
           }
         }
 
     assertEquals(BigInteger.ONE, testTSC.countAllPossibleNWayPredicateCombinations(2))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(2).all { it.elements.size == 2 },
+    )
+    assertEquals(1, testTSC.getAllPossibleNWayPredicateCombinations(2).size)
   }
 
   /** Test 2-way feature combination counting for any node. */
@@ -163,12 +292,16 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           any("root") {
-            leaf("leaf 1")
-            leaf("leaf 2")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
           }
         }
 
     assertEquals(BigInteger.ONE, testTSC.countAllPossibleNWayPredicateCombinations(2))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(2).all { it.elements.size == 2 },
+    )
+    assertEquals(1, testTSC.getAllPossibleNWayPredicateCombinations(2).size)
   }
 
   /** Test 2-way feature combination counting for exclusive node. */
@@ -177,12 +310,16 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           exclusive("root") {
-            leaf("leaf 1")
-            leaf("leaf 2")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
           }
         }
 
     assertEquals(BigInteger.ZERO, testTSC.countAllPossibleNWayPredicateCombinations(2))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(2).all { it.elements.size == 2 },
+    )
+    assertEquals(0, testTSC.getAllPossibleNWayPredicateCombinations(2).size)
   }
 
   /** Test 2-way feature combination counting for bounded node. */
@@ -191,12 +328,106 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           bounded("root", 0 to 2) {
-            leaf("leaf 1")
-            leaf("leaf 2")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
           }
         }
 
     assertEquals(BigInteger.ONE, testTSC.countAllPossibleNWayPredicateCombinations(2))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(2).all { it.elements.size == 2 },
+    )
+    assertEquals(1, testTSC.getAllPossibleNWayPredicateCombinations(2).size)
+  }
+
+  /** Test 2-way feature combination counting for with inner feature nodes. */
+  @Test
+  fun `Test 2-way feature combination counting for with inner feature nodes without condition`() {
+    val testTSC =
+        tsc<E, T, S, U, D> {
+          all("root") {
+            optional("optional 1") {
+              optional("optional 2") { leaf("leaf 1") { condition { true } } }
+            }
+          }
+        }
+
+    assertEquals(BigInteger.ZERO, testTSC.countAllPossibleNWayPredicateCombinations(2))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(2).all { it.elements.size == 2 },
+    )
+    assertEquals(0, testTSC.getAllPossibleNWayPredicateCombinations(2).size)
+  }
+
+  /** Test 2-way feature combination counting with one inner feature node with a condition. */
+  @Test
+  fun `Test 2-way feature combination counting with one inner feature node with a condition`() {
+    val testTSC =
+        tsc<E, T, S, U, D> {
+          all("root") {
+            optional("optional 1") {
+              condition { true }
+              optional("optional 2") { leaf("leaf 1") { condition { true } } }
+            }
+          }
+        }
+
+    assertEquals(1, testTSC.getAllPossibleNWayPredicateCombinations(2).size)
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(2).all { it.elements.size == 2 },
+    )
+    assertEquals(BigInteger.ONE, testTSC.countAllPossibleNWayPredicateCombinations(2))
+  }
+
+  /** Test 2-way feature combination counting with two inner feature nodes with a condition. */
+  @Test
+  fun `Test 2-way feature combination counting with two inner feature nodes with a condition`() {
+    val testTSC =
+        tsc<E, T, S, U, D> {
+          all("root") {
+            optional("optional 1") {
+              condition { true }
+              optional("optional 2") {
+                condition { true }
+                leaf("leaf 1") { condition { true } }
+              }
+            }
+          }
+        }
+
+    assertEquals(3, testTSC.getAllPossibleNWayPredicateCombinations(2).size)
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(2).all { it.elements.size == 2 },
+    )
+    assertEquals(BigInteger("3"), testTSC.countAllPossibleNWayPredicateCombinations(2))
+  }
+
+  /** Test 2-way feature combination counting with two inner feature nodes with a condition. */
+  @Test
+  fun `Test 2-way feature combination counting with two inner feature nodes without a condition`() {
+    val testTSC =
+        tsc<E, T, S, U, D> {
+          all("root") {
+            optional("optional 1") {
+              optional("optional 2") { leaf("leaf 1") { condition { true } } }
+            }
+          }
+        }
+
+    assertEquals(0, testTSC.getAllPossibleNWayPredicateCombinations(2).size)
+    assertEquals(BigInteger("0"), testTSC.countAllPossibleNWayPredicateCombinations(2))
+  }
+
+  /** Test 2-way feature combination counting with all nodes without a condition. */
+  @Test
+  fun `Test 2-way feature combination counting with all nodes without a condition`() {
+    val testTSC =
+        tsc<E, T, S, U, D> {
+          all("root") { optional("optional 1") { optional("optional 2") { leaf("leaf 1") } } }
+        }
+
+    assertEquals(0, testTSC.getAllPossibleNWayPredicateCombinations(2).size)
+    assertEquals(BigInteger("0"), testTSC.countAllPossibleNWayPredicateCombinations(2))
   }
 
   // endregion
@@ -209,14 +440,18 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           all("root") {
-            leaf("leaf 1")
-            leaf("leaf 2")
-            leaf("leaf 3")
-            leaf("leaf 4")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
+            leaf("leaf 3") { condition { true } }
+            leaf("leaf 4") { condition { true } }
           }
         }
 
     assertEquals(BigInteger("6"), testTSC.countAllPossibleNWayPredicateCombinations(2))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(2).all { it.elements.size == 2 },
+    )
+    assertEquals(6, testTSC.getAllPossibleNWayPredicateCombinations(2).size)
   }
 
   /** Test 2-way feature combination counting for optional node with larger TSC. */
@@ -225,14 +460,18 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           optional("root") {
-            leaf("leaf 1")
-            leaf("leaf 2")
-            leaf("leaf 3")
-            leaf("leaf 4")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
+            leaf("leaf 3") { condition { true } }
+            leaf("leaf 4") { condition { true } }
           }
         }
 
     assertEquals(BigInteger("6"), testTSC.countAllPossibleNWayPredicateCombinations(2))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(2).all { it.elements.size == 2 },
+    )
+    assertEquals(6, testTSC.getAllPossibleNWayPredicateCombinations(2).size)
   }
 
   /** Test 2-way feature combination counting for any node with larger TSC. */
@@ -241,14 +480,18 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           any("root") {
-            leaf("leaf 1")
-            leaf("leaf 2")
-            leaf("leaf 3")
-            leaf("leaf 4")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
+            leaf("leaf 3") { condition { true } }
+            leaf("leaf 4") { condition { true } }
           }
         }
 
     assertEquals(BigInteger("6"), testTSC.countAllPossibleNWayPredicateCombinations(2))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(2).all { it.elements.size == 2 },
+    )
+    assertEquals(6, testTSC.getAllPossibleNWayPredicateCombinations(2).size)
   }
 
   /** Test 2-way feature combination counting for exclusive node with larger TSC. */
@@ -257,14 +500,18 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           exclusive("root") {
-            leaf("leaf 1")
-            leaf("leaf 2")
-            leaf("leaf 3")
-            leaf("leaf 4")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
+            leaf("leaf 3") { condition { true } }
+            leaf("leaf 4") { condition { true } }
           }
         }
 
     assertEquals(BigInteger.ZERO, testTSC.countAllPossibleNWayPredicateCombinations(2))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(2).all { it.elements.size == 2 },
+    )
+    assertEquals(0, testTSC.getAllPossibleNWayPredicateCombinations(2).size)
   }
 
   /** Test 2-way feature combination counting for bounded node with larger TSC. */
@@ -273,14 +520,18 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           bounded("root", 1 to 3) {
-            leaf("leaf 1")
-            leaf("leaf 2")
-            leaf("leaf 3")
-            leaf("leaf 4")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
+            leaf("leaf 3") { condition { true } }
+            leaf("leaf 4") { condition { true } }
           }
         }
 
     assertEquals(BigInteger("6"), testTSC.countAllPossibleNWayPredicateCombinations(2))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(2).all { it.elements.size == 2 },
+    )
+    assertEquals(6, testTSC.getAllPossibleNWayPredicateCombinations(2).size)
   }
 
   // endregion
@@ -293,20 +544,24 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           all("root") {
-            leaf("leaf 1")
-            leaf("leaf 2")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
             optional("optional 1") {
-              leaf("leaf 3")
-              leaf("leaf 4")
+              leaf("leaf 3") { condition { true } }
+              leaf("leaf 4") { condition { true } }
               exclusive("exclusive 1") {
-                leaf("leaf 5")
-                leaf("leaf 6")
+                leaf("leaf 5") { condition { true } }
+                leaf("leaf 6") { condition { false } }
               }
             }
           }
         }
 
     assertEquals(BigInteger("14"), testTSC.countAllPossibleNWayPredicateCombinations(2))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(2).all { it.elements.size == 2 },
+    )
+    assertEquals(14, testTSC.getAllPossibleNWayPredicateCombinations(2).size)
   }
 
   // endregion
@@ -319,14 +574,18 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           all("root") {
-            leaf("leaf 1")
-            leaf("leaf 2")
-            leaf("leaf 3")
-            leaf("leaf 4")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
+            leaf("leaf 3") { condition { true } }
+            leaf("leaf 4") { condition { true } }
           }
         }
 
     assertEquals(BigInteger("4"), testTSC.countAllPossibleNWayPredicateCombinations(3))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(3).all { it.elements.size == 3 },
+    )
+    assertEquals(4, testTSC.getAllPossibleNWayPredicateCombinations(3).size)
   }
 
   /** Test 3-way feature combination counting for optional node with larger TSC. */
@@ -335,14 +594,18 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           optional("root") {
-            leaf("leaf 1")
-            leaf("leaf 2")
-            leaf("leaf 3")
-            leaf("leaf 4")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
+            leaf("leaf 3") { condition { true } }
+            leaf("leaf 4") { condition { true } }
           }
         }
 
     assertEquals(BigInteger("4"), testTSC.countAllPossibleNWayPredicateCombinations(3))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(3).all { it.elements.size == 3 },
+    )
+    assertEquals(4, testTSC.getAllPossibleNWayPredicateCombinations(3).size)
   }
 
   /** Test 3-way feature combination counting for any node with larger TSC. */
@@ -351,14 +614,18 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           any("root") {
-            leaf("leaf 1")
-            leaf("leaf 2")
-            leaf("leaf 3")
-            leaf("leaf 4")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
+            leaf("leaf 3") { condition { true } }
+            leaf("leaf 4") { condition { true } }
           }
         }
 
     assertEquals(BigInteger("4"), testTSC.countAllPossibleNWayPredicateCombinations(3))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(3).all { it.elements.size == 3 },
+    )
+    assertEquals(4, testTSC.getAllPossibleNWayPredicateCombinations(3).size)
   }
 
   /** Test 3-way feature combination counting for exclusive node with larger TSC. */
@@ -367,14 +634,18 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           exclusive("root") {
-            leaf("leaf 1")
-            leaf("leaf 2")
-            leaf("leaf 3")
-            leaf("leaf 4")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
+            leaf("leaf 3") { condition { true } }
+            leaf("leaf 4") { condition { true } }
           }
         }
 
     assertEquals(BigInteger.ZERO, testTSC.countAllPossibleNWayPredicateCombinations(3))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(3).all { it.elements.size == 3 },
+    )
+    assertEquals(0, testTSC.getAllPossibleNWayPredicateCombinations(3).size)
   }
 
   /** Test 3-way feature combination counting for bounded node with larger TSC. */
@@ -383,14 +654,18 @@ class TSCNWayFeatureCombinationCountingTest {
     val testTSC =
         tsc<E, T, S, U, D> {
           bounded("root", 1 to 3) {
-            leaf("leaf 1")
-            leaf("leaf 2")
-            leaf("leaf 3")
-            leaf("leaf 4")
+            leaf("leaf 1") { condition { true } }
+            leaf("leaf 2") { condition { true } }
+            leaf("leaf 3") { condition { true } }
+            leaf("leaf 4") { condition { true } }
           }
         }
 
     assertEquals(BigInteger("4"), testTSC.countAllPossibleNWayPredicateCombinations(3))
+    assertTrue(
+        testTSC.getAllPossibleNWayPredicateCombinations(3).all { it.elements.size == 3 },
+    )
+    assertEquals(4, testTSC.getAllPossibleNWayPredicateCombinations(3).size)
   }
 
   // endregion
@@ -400,7 +675,10 @@ class TSCNWayFeatureCombinationCountingTest {
   /** Test 3-way feature combination counting for all node with huge TSC (20,000 leaf nodes). */
   @Test
   fun `Test 3-way feature combination counting for all node with huge TSC (20,000 leaf nodes)`() {
-    val testTSC = tsc<E, T, S, U, D> { all("root") { repeat(20_000) { leaf("leaf $it") } } }
+    val testTSC =
+        tsc<E, T, S, U, D> {
+          all("root") { repeat(20_000) { leaf("leaf $it") { condition { true } } } }
+        }
 
     assertEquals(BigInteger("1333133340000"), testTSC.countAllPossibleNWayPredicateCombinations(3))
   }
@@ -410,7 +688,10 @@ class TSCNWayFeatureCombinationCountingTest {
    */
   @Test
   fun `Test 3-way feature combination counting for optional node with huge TSC (20,000 leaf nodes)`() {
-    val testTSC = tsc<E, T, S, U, D> { optional("root") { repeat(20_000) { leaf("leaf $it") } } }
+    val testTSC =
+        tsc<E, T, S, U, D> {
+          optional("root") { repeat(20_000) { leaf("leaf $it") { condition { true } } } }
+        }
 
     assertEquals(BigInteger("1333133340000"), testTSC.countAllPossibleNWayPredicateCombinations(3))
   }
@@ -418,7 +699,10 @@ class TSCNWayFeatureCombinationCountingTest {
   /** Test 3-way feature combination counting for any node with huge TSC (20,000 leaf nodes). */
   @Test
   fun `Test 3-way feature combination counting for any node with huge TSC (20,000 leaf nodes)`() {
-    val testTSC = tsc<E, T, S, U, D> { any("root") { repeat(20_000) { leaf("leaf $it") } } }
+    val testTSC =
+        tsc<E, T, S, U, D> {
+          any("root") { repeat(20_000) { leaf("leaf $it") { condition { true } } } }
+        }
 
     assertEquals(BigInteger("1333133340000"), testTSC.countAllPossibleNWayPredicateCombinations(3))
   }
@@ -428,7 +712,10 @@ class TSCNWayFeatureCombinationCountingTest {
    */
   @Test
   fun `Test 3-way feature combination counting for exclusive node with huge TSC (20,000 leaf nodes)`() {
-    val testTSC = tsc<E, T, S, U, D> { exclusive("root") { repeat(20_000) { leaf("leaf $it") } } }
+    val testTSC =
+        tsc<E, T, S, U, D> {
+          exclusive("root") { repeat(20_000) { leaf("leaf $it") { condition { true } } } }
+        }
 
     assertEquals(BigInteger.ZERO, testTSC.countAllPossibleNWayPredicateCombinations(3))
   }
@@ -437,7 +724,9 @@ class TSCNWayFeatureCombinationCountingTest {
   @Test
   fun `Test 3-way feature combination counting for bounded node with huge TSC (20,000 leaf nodes)`() {
     val testTSC =
-        tsc<E, T, S, U, D> { bounded("root", 1 to 3) { repeat(20_000) { leaf("leaf $it") } } }
+        tsc<E, T, S, U, D> {
+          bounded("root", 1 to 3) { repeat(20_000) { leaf("leaf $it") { condition { true } } } }
+        }
 
     assertEquals(BigInteger("1333133340000"), testTSC.countAllPossibleNWayPredicateCombinations(3))
   }