+++ CHANGELOG.md
@@ -34,6 +34,11 @@ All notable changes to this project will be documented in this file.
 - Add `NWayFeatureCombinationsPerTSCMetric`.
 - Add `identifier` field to `MetricProvider` and allow the registration of multiple instances with the same `MetricProvider` class when the `identifiers` differ.
 - Add performant `isValid()` function to `TSCInstance` and `TSCInstanceNode`.
+- Add `TSC.countAllPossibleNWayPredicateCombinations()` and `TSC.getAllPossibleNWayPredicateCombinations()`functions.
+
+### Changed
+
+- Rename `TSCInstanceNode.getLeafNodeEdges()` to `TSCInstanceNode.getFeatureNodeEdges()`. 
 
 
 ## [1.0] - 14.10.2025